+++ src/components/Accordion/index.js
@@ -1,2 +1,3 @@
 export * from './Accordion';
 export * from './AccordionArticle';
+export * from './AppAccordionArticle';